+++ src/platform/qt/utils.h
@@ -31,7 +31,7 @@ enum class Endian {
 };
 
 QString niceSizeFormat(size_t filesize);
-QString nicePlatformFormat(mPlatform platform, uint8_t validModels = 0);
+QString nicePlatformFormat(mPlatform platform, int validModels = 0);
 
 bool convertAddress(const QHostAddress* input, Address* output);
